+++ src/Pattern.ts
